backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "fastapi>=0.104.0",
    "httpx>=0.20",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.4.2",
    "pydantic-settings>=2.0.3",
    "pytest-asyncio>=0.15",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "sqlmodel>=0.0.11",
    "supabase>=2.15.3",
    "uvicorn>=0.24.0",
]
